Use sinf() instead of sin() for sample generation to reduce computational overhead somewhat

+++ src/MMToneGenerator.m
@@ -46,7 +46,7 @@
 	{
 		float sample = 0;
 		for ( unsigned j=0; j<numTones; ++j )
-			sample += amplitudes[j] * sin( (offset + i) * multipliers[j] );
+			sample += amplitudes[j] * sinf( (offset + i) * multipliers[j] );
 		samples[i] += roundf( sample );
 	}
 }
